Stretch: Creates bears table and adds all CRUD endpoints
