+++ .gitIgnore
@@ -62,3 +62,5 @@ typings/
 
 # VSCode settings
 .vscode/
+
+node_modules
